_packages/navee_utils/requirements.txt
boto3==1.25.1
imagehash==4.3.1
numpy==1.24.2
pandas>=1.1.4
pre-commit==2.16.0
Pillow==9.5.0
pillow-avif-plugin==1.3.1
pytest==5.3.5
pytest-dependency==0.5.1
requests==2.28.1
tldextract==3.4.4
urllib3==1.26.12